Working sketch of HIE->neo4j direction
Shoved most single-file AST information into Neo4J, including the space-
efficient type representation in HIE.

The codebase is a little messy right now because of the ad-hoc treatment
of BoltActionT vs. Free Expr () -- between the query action and the
DSL monad resp. It's surprising sometimes where the query builder needs
to cross boundaries, for example when preserving a node identifier (the
temporary name given to nodes in queries) I need to encapsulate all the
dependents into the same query. This is currently a bit dangerous given
the maximum default query length is ~50Kchar, given I shove all the
properties of an AST node in a single query so I can point the node to
all its non-flat/normalized properties (e.g. NodeContext information).
The more reasonable strategy may be to return IDs from node-forming
queries and string them along BoltActionT, although I don't love the
error-prone pattern matching to extract the ID from the result sets,
if even for how verbose and imperative it is. Unfortunately there is
no translation from a HasBOLT node Structure object to a Cypher node
AFAIK, so we can't sub in a node straight into the next query when
it's returned from the previous one. ID's seem to be the Right Way to
do this.

Many of the reusable function definitions are also ugly and ad hoc,
often unique'd by a filepath that it closes over and with conflicting
requirements to finalize into a BoltActionT or keep as in the DSL
(compare the methods in types2cypher vs. ast2cypher).

Finally, one last bit of ugliness is the ordering of queries. Because
much of the relationship construction is done without much information
about at least one of the nodes (e.g. in types, we only know the type
ID), we either have to insert all of the nodes first, or pre-empt with
an incomplete node and fill in the properties of that later, which
turns nice MERGE clauses into uglier SET clauses and may leave
incomplete nodes if we never get to filling them out/if the
relationship is ever invalid. In types2cypher, I have to iterate over
the list of types twice explicitly, which is kinda upsetting.
Initially I wanted to wrap an appropriate monad/transformer to sort the
queries so the node queries happen first, but the explicit refactor is
so much more straightforward and my ordering rules wouldn't be that
elegant anyways (e.g. merge vs. match).