Connect AST nodes to children + named nodes via trigger
The name trigger is a bit coarse, I should namespace my `name` field
or restrict it to the labels that are associated with graphie.

Still, this is the first incarnation of the schema that seems to have
most of what we want, so development will start on the frontend now.