Format import command
Test graphie on its own dependences (and itself) by spitting out the bash
command to import all the nodes properly.

Realizing that nodes may want more infomation depending on how deep the
tree is and the search strategy. In particular, we may need to include
file name under each node, which motivates the reintroduction of the
filename name stash (although yet to be seen is whether this is that much
more space-efficeint, shoving N more relationships into the graph)